Correct ZipProgressEventArgs to use Int64 for BytesTransferred.  Also changes in the WinForms example app - specifically around the progress bar. Also, a small update to enable DotNetZip to read WinZip SFX archives. And some tests to go with that. Last, use unchecked() in the CRC32 class when casting from uint to int.
git-svn-id: https://dotnetzip.svn.codeplex.com/svn@28829 4ef1bfeb-0a34-4d02-9305-240e5b07a168

+++ Examples/WinFormsApp/Form1.cs
@@ -384,7 +384,8 @@ namespace WinFormsExample
                             _nFilesCompleted + 1, _entriesToZip, e.CurrentEntry.FileName);
                     }
 
-                    int xferred = e.BytesTransferred >> _progress2MaxFactor;
+                    // downcast should be safe because we have shifted e.BytesTransferred
+                    int xferred = (int) (e.BytesTransferred >> _progress2MaxFactor);
 
                     this.progressBar2.Value = (xferred >= this.progressBar2.Maximum)
                         ? this.progressBar2.Maximum
